gwssrv: pixel

+++ gramado/core/aurora/server/view/pixel.h
@@ -9,23 +9,23 @@
 int 
 pixelBackBufferPutpixel2 ( 
     unsigned long color, 
-    unsigned long x, 
-    unsigned long y );
+    int x, 
+    int y );
 
 
 //ACESSANDO O BACKBUFFER DIRETAMENTE
 int 
 pixelBackBufferPutpixel ( 
     unsigned long color, 
-    unsigned long x, 
-    unsigned long y );
+    int x, 
+    int y );
     
 
 
 unsigned long
 pixelBackBufferGetColor ( 
-    unsigned long x, 
-    unsigned long y );
+    int x, 
+    int y );
     
 
 #endif
